requirements.txt
djangorestframework==3.14.0
opencv_python==4.7.0.72
orjson==3.9.1
plotly==5.14.0
azure-storage-blob==12.16.0
asgiref==3.7.2
contourpy==1.1.0
cycler==0.11.0
Django==4.2.3
fonttools==4.41.1
kiwisolver==1.4.4
matplotlib==3.7.2
numpy==1.23.5
packaging==23.1
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.4
